fix lammps tests and cleanup the module
